refactor: add Section Item component
