+++ src/components/showMore/Section/styled.ts
@@ -1,8 +1,4 @@
-import { Animated } from 'react-native';
-
 import styled from '@emotion/native';
-
-import { colors } from 'src/styles';
 
 export const SectionContainer = styled.View`
   width: 100%;
@@ -15,23 +11,4 @@ export const SectionContainer = styled.View`
 export const Section = styled.View`
   flex-direction: column;
   row-gap: 20px;
-`;
-
-export const SectionItem = styled(Animated.View)`
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  border-radius: 16px;
-`;
-
-export const SectionIconContainer = styled.View`
-  flex-direction: row;
-  align-items: center;
-  column-gap: 10px;
-`;
-
-export const SectionIconWrapper = styled.View`
-  padding: 4px;
-  border-radius: 10px;
-  background-color: ${colors.lightGray};
 `;
